Conflicts:
	src/main/java/egovframework/ezEKP/ezAttitude/web/EzAttitudeGWController.java
	src/main/webapp/WEB-INF/jsp/ezAttitude/attitudeUserMain.jsp

+++ src/main/java/egovframework/ezEKP/ezOrgan/web/EzOrganAdminController.java
@@ -1,11 +1,11 @@
 package egovframework.ezEKP.ezOrgan.web;
 
+import java.awt.Color;
 import java.awt.Graphics2D;
 import java.awt.geom.Ellipse2D;
 import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.FileOutputStream;
-import java.io.FileReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -13,8 +13,10 @@ import java.nio.file.Files;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
+import java.util.Map;
 import java.util.Properties;
 import java.util.TimeZone;
 import java.util.UUID;
@@ -24,10 +26,7 @@ import javax.imageio.ImageIO;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import net.minidev.json.JSONArray;
-import net.minidev.json.JSONObject;
-import net.minidev.json.parser.JSONParser;
-
+import org.json.simple.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -36,8 +35,6 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.multipart.MultipartHttpServletRequest;
@@ -58,7 +55,6 @@ import egovframework.ezEKP.ezOrgan.service.EzOrganService;
 import egovframework.ezEKP.ezOrgan.vo.OrganDeptVO;
 import egovframework.ezEKP.ezOrgan.vo.OrganUserVO;
 import egovframework.ezEKP.ezSystem.service.EzSystemAdminService;
-import egovframework.ezEKP.ezSystem.vo.ConnectionInfoVO;
 import egovframework.let.user.login.vo.LoginSimpleVO;
 import egovframework.let.user.login.vo.LoginVO;
 import egovframework.let.utl.fcc.service.ClientUtil;
@@ -137,7 +133,16 @@ public class EzOrganAdminController extends EgovFileMngUtil {
 		LoginVO user = commonUtil.userInfo(loginCookie);
 		String dotNetIntegration = ezCommonService.getTenantConfig("dotNetIntegration", user.getTenantId());
 		
+		// set useLetter
+		String useLetter = ezCommonService.getTenantConfig("useLetter", user.getTenantId());
+		if (useLetter == null || useLetter.equals("")) {
+			useLetter = "NO";
+		}
+				
+		logger.debug("useLetter=" + useLetter);
+		
 		model.addAttribute("dotNetIntegration", dotNetIntegration);
+		model.addAttribute("useLetter", useLetter);
 		
 		return "admin/ezOrgan/organLeft";
 	}
@@ -282,18 +287,32 @@ public class EzOrganAdminController extends EgovFileMngUtil {
 					
 					rc = ezEmailUserAdminService.updateGroupAdd(groupAddr, mailAddr);
 					
+					//업무일지 - 일지함 생성
+					Map<String, Object> param = new HashMap<String, Object>();
+					param.put("tenantId", tenantID);
+					param.put("companyId", cn);
+					param.put("userId", userInfo.getId());
+					JSONObject journalResult = commonUtil.getJsonFromRestApi("/rest/ezjournal/types", param, request, "post", null);
+					
+					String journalStatus = (String) journalResult.get("status");
+					
+					if (journalStatus.equals("ok")) {
+						ezEmailUserAdminService.updateGroupDel(groupAddr, mailAddr);
+					}
+					
 					logger.debug("updateGroupAdd rc=" + rc);
 					
-					if (rc == 0) { // updateGroupAdd 성공
+					if (rc == 0 && journalStatus.equals("ok")) { // updateGroupAdd 성공
 						
 						// insertDBData_company 실패했을 경우 JMocha에서 회사 다시 삭제.
 						try {
 							ezOrganAdminService.insertDBData_company(cn, displayName, displayName2,
 									mailAddr, parentCn, ldapPath, extensionAttribute15, skipInitData, tenantID, userInfo);
-							result = "OK";	
+							result = "OK";
 						} catch (Exception e) {
 							e.printStackTrace();
-							
+							commonUtil.getJsonFromRestApi("/rest/ezjournal/types", param, request, "delete", null);
+
 							ezEmailUserAdminService.updateGroupDel(groupAddr, mailAddr);
 							ezEmailUserAdminService.removeGroup(mailAddr);
 							result = "EMAIL_ERROR";
@@ -1268,7 +1287,7 @@ public class EzOrganAdminController extends EgovFileMngUtil {
 	        ezOrganAdminService.updateDBData_user(vo);
 	        result = "OK";
 		// 새로운 사용자를 등록한다.
-		} else {		    
+		} else {
 			String domain = ezCommonService.getTenantConfig("DomainName", tenantID);
 			String cn = vo.getCn();
 						
@@ -1607,10 +1626,14 @@ public class EzOrganAdminController extends EgovFileMngUtil {
 
 			writeUploadedFile(multiFile, fileName + extension, tempPath);
 			File imageFile = new File(tempPath + fileName + extension);			
-			
+
 			BufferedImage bi = ImageIO.read(imageFile);
-            BufferedImage bufferedImage = new BufferedImage(119, 128, bi.getType());
-            bufferedImage.createGraphics().drawImage(bi, 0, 0, 119, 128, null);
+			/*2018-04-12이효진  bi.getType으로 지정시 color변경되어 TYPE_4BYTE_ABGR로 지정*/
+//            BufferedImage bufferedImage = new BufferedImage(119, 128, bi.getType());
+            BufferedImage bufferedImage = new BufferedImage(119, 128, BufferedImage.TYPE_4BYTE_ABGR);
+            /*2018-04-12이효진  PNG파일 배경지정*/
+//            bufferedImage.createGraphics().drawImage(bi, 0, 0, 119, 128, null);
+            bufferedImage.createGraphics().drawImage(bi, 0, 0, 119, 128, Color.WHITE, null);
             
             File file2 = new File(serverPath + fileName + "png");
             ImageIO.write(bufferedImage, "png", file2);
@@ -2432,10 +2455,11 @@ public class EzOrganAdminController extends EgovFileMngUtil {
 			if (userInfo.getRollInfo().indexOf("c=1") == -1 && userInfo.getRollInfo().indexOf("k=1") == -1) {
 				return returnValue;
 			}
+			int tenantId = userInfo.getTenantId();
 			
 			Document xmldom = commonUtil.convertStringToDocument(bodyData);
 			String mail = xmldom.getElementsByTagName("MAIL").item(0).getTextContent();
-			returnValue = ezEmailService.checkIndividualAlias(mail);
+			returnValue = ezEmailService.checkIndividualAlias(mail,tenantId);
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
